% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{build_caches}
\alias{build_caches}
\title{Build local disk caches}
\usage{
build_caches(
  GeVIR = TRUE,
  HGNC = TRUE,
  GTEx = TRUE,
  IGV = TRUE,
  UCSC = TRUE,
  PanelApp = TRUE,
  HPO = TRUE,
  MI_OMIM = TRUE,
  Genes4Epilepsy = TRUE
)
}
\description{
Runs all functions that use local disk caching to build a complete cache for running in offline or fallback mode
}
