% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelApp.R
\name{build_panelapp_cache}
\alias{build_panelapp_cache}
\title{Store local cache of all PanelApp panels}
\usage{
build_panelapp_cache(sources = names(get_cavalier_opt("panelapp_urls")))
}
\description{
Store local cache of all PanelApp panels
}
