% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache}
\alias{cache}
\title{Save or load disk cache}
\usage{
cache(
  name,
  fun = NULL,
  version = NULL,
  subdir = NULL,
  cache_dirs = get_cache_dirs()
)
}
\description{
Store result of fun() in RDS file on disk, or load from disk if it exists
will store files in one of:
   <cache_dir>/<name>.rds
   <cache_dir>/<name>.ver_<version>.rds
   <cache_dir>/<subdir>/<name>.rds
   <cache_dir>/<subdir>/<name>.ver_<version>.rds
}
