% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_file}
\alias{cache_file}
\title{Store or load extenal file to disk}
\usage{
cache_file(url, version = NULL, subdir = NULL, cache_dirs = get_cache_dirs())
}
\description{
Store or load extenal file to disk
}
