% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavalier.R
\docType{package}
\name{cavalier}
\alias{cavalier-package}
\alias{cavalier}
\title{cavalier: Candidate Variant List Evaluation Report}
\description{
\pkg{cavalier} assists and automates variant interpretation in next-generation sequencing data:
1) filter list of variants and create a short list of candidate variants
2) combine information from multiple sources useful for evaluating and prioritising variants
3) create PDF output for visualing and communicating results
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bahlolab/cavalier}
  \item Report bugs at \url{https://github.com/bahlolab/cavalier/issues}
}

}
\author{
\strong{Maintainer}: Mark Bennett \email{mark.bennett@wehi.edu.au}

Authors:
\itemize{
  \item Jacob Munro \email{munro.j@wehi.edu.au}
  \item Neblina Sikta \email{sikta.n@wehi.edu.au}
}

}
