% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_memory_caches}
\alias{clear_memory_caches}
\title{Removes memory cache for functions memoised with memoise::memoise}
\usage{
clear_memory_caches()
}
\description{
Removes memory cache for functions memoised with memoise::memoise
}
