% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_intolerance.R
\name{ensembl2loeuf}
\alias{ensembl2loeuf}
\title{Convert ensemble gene id to to (ExAC) LOUEF percentile}
\usage{
ensembl2loeuf(ensembl)
}
\description{
Convert ensemble gene id to to (ExAC) LOUEF percentile
}
