% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UCSC.R
\name{get_centromeres_gaps}
\alias{get_centromeres_gaps}
\title{Get assembly gap and centromere locations for get_cavalier_opt('ref_genome')}
\usage{
get_centromeres_gaps(ref_genome = get_cavalier_opt("ref_genome"))
}
\description{
Get assembly gap and centromere locations for get_cavalier_opt('ref_genome')
}
