% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPO.R
\name{get_disease_names}
\alias{get_disease_names}
\title{Convert HPO disease_id to disease name}
\usage{
get_disease_names(disease_ids, db_mode = get_cavalier_opt("database_mode"))
}
\description{
Uses HPO API or disk cache depending on cavalier_opts$database_mode
}
