% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genes4Epilepsy.R
\name{get_g4e_full_list}
\alias{get_g4e_full_list}
\title{Retrieve Genes4Epilepsy from GitHub}
\usage{
get_g4e_full_list(version = get_g4e_version())
}
\description{
https://github.com/bahlolab/Genes4Epilepsy
}
