% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genes4Epilepsy.R
\name{get_g4e_phenotype_list}
\alias{get_g4e_phenotype_list}
\title{Get sublist of Genes4Epilepsy by phenotype}
\usage{
get_g4e_phenotype_list(
  version = get_g4e_version(),
  phenotype = c("Epilepsy", "DEE", "Focal", "MCD", "GGE", "PME")
)
}
\description{
Get sublist of Genes4Epilepsy by phenotype
}
