% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genes4Epilepsy.R
\name{get_g4e_version}
\alias{get_g4e_version}
\title{Get HPO version from GitHub or disk cache}
\usage{
get_g4e_version(db_mode = get_cavalier_opt("database_mode"))
}
\description{
Get HPO version from GitHub or disk cache
}
