% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{get_gencode_coords}
\alias{get_gencode_coords}
\title{Get HGNC complete table from either get_cavalier_opt("hgnc_monthly_base_url") or disk cache}
\usage{
get_gencode_coords(ver = get_gencode_version())
}
\description{
Get HGNC complete table from either get_cavalier_opt("hgnc_monthly_base_url") or disk cache
}
