% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPO.R
\name{get_gene_disease_map}
\alias{get_gene_disease_map}
\title{Simplified mapping from gene to phenotype from HPO}
\usage{
get_gene_disease_map(source = c("ALL", "OMIM", "ORPHA"))
}
\description{
Simplified mapping from gene to phenotype from HPO
}
