% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_list.R
\name{get_gene_list}
\alias{get_gene_list}
\title{Return a gene list from PanelApp, HPO or Genes4Epilepsy}
\usage{
get_gene_list(id, version = NULL, save = NULL, min_confidence = 2L)
}
\description{
Return data frame with columns list_id, list_name, list_version, symbol, hgnc_id, ensembl_gene_id, entrez_id, inheritance
NB confidence_level only used for PanelApp.
}
\examples{
# Get PanelApp Australia Panel 202 (Genetic Epilespsy)
get_gene_list('PAA:202', version = '1.26')
# Get HPO genes associated with HP:0001250 (Seizure)
get_gene_list('HP:0001250')
#' Get all Genes4Epilepsy genes
get_gene_list('G4E:Epilepsy')
# Get Genes4Epilepsy genes associated Malformations of Cortical Development
get_gene_list('G4E:MCD')
}
