% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_intolerance.R
\name{get_gevir_table}
\alias{get_gevir_table}
\title{Retrieve GeVIR table from get_cavalier_opt("gevir_url") or disk cache}
\usage{
get_gevir_table()
}
\description{
Retrieve GeVIR table from get_cavalier_opt("gevir_url") or disk cache
}
