% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_intolerance.R
\name{get_gevir_table_hgnc}
\alias{get_gevir_table_hgnc}
\title{Update GeVIR table gene symbol with current HGNC symbol}
\usage{
get_gevir_table_hgnc()
}
\description{
Update GeVIR table gene symbol with current HGNC symbol
}
