% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx.R
\name{get_gtex_expression_hgnc}
\alias{get_gtex_expression_hgnc}
\title{Replace symbol in GTEx expression with current HGNC version}
\usage{
get_gtex_expression_hgnc()
}
\description{
Replace symbol in GTEx expression with current HGNC version
}
