% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC.R
\name{get_hgnc_alias}
\alias{get_hgnc_alias}
\title{Get HGNC gene symbols alias table to convert symbols to current HGNC symbol}
\usage{
get_hgnc_alias(ver = get_hgnc_version())
}
\description{
Get HGNC gene symbols alias table to convert symbols to current HGNC symbol
}
