% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC.R
\name{get_hgnc_complete}
\alias{get_hgnc_complete}
\title{Get HGNC complete table from either get_cavalier_opt("hgnc_monthly_base_url") or disk cache}
\usage{
get_hgnc_complete(
  ver = get_cavalier_opt("hgnc_ver"),
  local_file = get_cavalier_opt("hgnc_local_file")
)
}
\description{
Get HGNC complete table from either get_cavalier_opt("hgnc_monthly_base_url") or disk cache
}
