% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC.R
\name{get_hgnc_locus_group_list}
\alias{get_hgnc_locus_group_list}
\title{Get list of from HGNC by locus_group}
\usage{
get_hgnc_locus_group_list(
  locus_group = c("protein-coding gene", "non-coding RNA", "pseudogene", "other", "ALL"),
  ver = get_hgnc_version()
)
}
\description{
Get list of from HGNC by locus_group
}
