% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HGNC.R
\name{get_hgnc_version}
\alias{get_hgnc_version}
\title{Get HGNC version from either get_cavalier_opt("hgnc_monthly_base_url") or disk cache}
\usage{
get_hgnc_version(db_mode = get_cavalier_opt("database_mode"))
}
\description{
Get HGNC version from either get_cavalier_opt("hgnc_monthly_base_url") or disk cache
}
