% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPO.R
\name{get_hpo_g2p_p2g}
\alias{get_hpo_g2p_p2g}
\title{Retrieve HPO genes_to_phenotype and phenotype_to_genes}
\usage{
get_hpo_g2p_p2g(hpo_version = get_hpo_version())
}
\description{
Download or load from disk cache
}
