% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPO.R
\name{get_hpo_term_names}
\alias{get_hpo_term_names}
\title{Mapping of hpo_term_id to hpo_term_name}
\usage{
get_hpo_term_names(hpo_version = get_hpo_version())
}
\description{
Mapping of hpo_term_id to hpo_term_name
}
