% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPO.R
\name{get_hpo_version}
\alias{get_hpo_version}
\title{Get HPO version from GitHub or disk cache}
\usage{
get_hpo_version(
  db_mode = get_cavalier_opt("database_mode"),
  ver = get_cavalier_opt("hpo_version")
)
}
\description{
Get HPO version from GitHub or disk cache
}
