% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelApp.R
\name{get_panelapp_panel_version}
\alias{get_panelapp_panel_version}
\title{Get latest version of PanelApp Panel (from server or cache)}
\usage{
get_panelapp_panel_version(id)
}
\description{
Get latest version of PanelApp Panel (from server or cache)
}
