% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_version}
\alias{get_version}
\title{Get the latest version from online on cache depending on database_mode}
\usage{
get_version(resource_name, cache_name, cache_subdir, func_online, db_mode)
}
\description{
Get the latest version from online on cache depending on database_mode
}
