% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPO.R
\name{hpo_api_get_disease_names}
\alias{hpo_api_get_disease_names}
\title{Query HPO API for disease name with disease id}
\usage{
hpo_api_get_disease_names(disease_ids)
}
\description{
Query HPO API for disease name with disease id
}
