% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPO.R
\name{hpo_api_get_diseases_by_gene}
\alias{hpo_api_get_diseases_by_gene}
\title{Query HPO API for diseases associates with gene, returning table of ids and name}
\usage{
hpo_api_get_diseases_by_gene(entrez_ids)
}
\description{
Query HPO API for diseases associates with gene, returning table of ids and name
}
