% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPO.R
\docType{data}
\name{hpo_mendelian_inheritnace}
\alias{hpo_mendelian_inheritnace}
\title{Mendelian inheritance terms for assigning inheritance to HPO diseases}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
hpo_mendelian_inheritnace
}
\description{
Mendelian inheritance terms for assigning inheritance to HPO diseases
}
\keyword{datasets}
