% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTEx.R
\name{plot_gtex_expression}
\alias{plot_gtex_expression}
\title{Plot GTEx tissue median RPKM expression for given gene symbol}
\usage{
plot_gtex_expression(gene, ensembl_id = NULL)
}
\description{
Plot GTEx tissue median RPKM expression for given gene symbol
}
